LPS-164308 Move object validation code to a new directory
