+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ObjectValidation/AddObjectValidation.tsx
@@ -160,7 +160,7 @@ function ModalAddObjectValidation({
 	);
 }
 
-export default function ModalWithProvider({
+export default function AddObjectValidation({
 	apiURL,
 	objectValidationRuleEngines,
 }: IProps) {
