+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/object_definitions/edit_object_validation.jsp
@@ -22,6 +22,6 @@ ObjectValidationRule objectValidationRule = (ObjectValidationRule)request.getAtt
 %>
 
 <react:component
-	module="js/components/EditObjectValidation"
+	module="js/components/ObjectValidation/EditObjectValidation"
 	props="<%= objectDefinitionsValidationsDisplayContext.getProps(objectValidationRule) %>"
 />
