+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ObjectValidation/AddObjectValidation.d.ts
@@ -14,7 +14,7 @@
 
 /// <reference types="react" />
 
-export default function ModalWithProvider({
+export default function AddObjectValidation({
 	apiURL,
 	objectValidationRuleEngines,
 }: IProps): JSX.Element;
