+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ObjectValidation/EditObjectValidation.d.ts
