+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ObjectValidation/ObjectValidationFormBase.d.ts
@@ -12,8 +12,6 @@
  * details.
  */
 
-/// <reference types="react" />
-
 import {FormError} from '@liferay/object-js-components-web';
 export declare function useObjectValidationForm({
 	initialValues,
